fix: fixes table of contents parsing
Updated README.md file to fix the table of contents parsing issue, the
heading should now be linkable, however, I can't say for sure that the
linking would always work. Trying to figure out if there is a correct
way to make linkable headings that trasfer to the correspondant section.

Before this commit, the table of contents were not parsed correctly.
The headings contained [Heading Title] (#link) displayed, when it should
only display a linkable "Heading Title"

Resolves: #1