docs: Fixes "Edit" section format
This should fix the "Edit" section formatting. "Edit" section now should
be wrapped inside a coding frame, contain the two terminal commands,
each is represented on a separate line.

Before this commit, the "Edit" section was a little confusing. It didn't
indicate whether the written line is a terminal command or not, besides,
it was representing the two separate commads as a single line.

Resolves: #2